phase 3 제출 전
