+++ Phase3_SEMO/src/dao/UserDAO.java
@@ -4,9 +4,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-
+import main.DBUtil;
 import dto.UserDTO;
 
 public class UserDAO {
