+++ Phase3_SEMO/src/page/Mypage.java
@@ -5,8 +5,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Scanner;
-
-import dao.DBUtil;
+import main.DBUtil;
 import dao.ThingDAO;
 import dao.ThingRankDAO;
 import dao.UserDAO;
@@ -20,6 +19,7 @@ public class Mypage extends UserPage {
 		this.setUser_id(my_id);
 	}
 	
+	// Mypage 메뉴를 보여줌
 	public void display(Scanner keyboard) {
 		int menu = 0;
 		while(true) {
@@ -68,6 +68,7 @@ public class Mypage extends UserPage {
 		
 	}
 	
+	//나(로그인한 user)에 대한 정보 보여줌
 	public void myInfo() {
 		try {
 			System.out.println(UserDAO.get_userInfo(this.getUser_id()).toString());
@@ -77,49 +78,20 @@ public class Mypage extends UserPage {
 		}
 	}
 	
-	// 카테고리 선택해서 카테고리 이름 반환하는 method
-	public String selectCategory(Scanner keyboard) {
-		System.out.println("choose category by number");
-		System.out.println("1. 한식\t\t2. 중식\t\t3. 일식/수산물\t4. 분식\t\t5. 닭/오리요리\n"
-				+ "6. 양식\t\t7. 패스트푸드\t8. 제과제빵떡케익\t9. 유흥주점\t10. 별식/퓨전요리\n"
-				+ "11. 커피점/카페\t12. 음식배달서비스\t13. 기타음식업\t14. 부페\t");
-		
-		int category_num = keyboard.nextInt();
-		
-		switch(category_num) {
-		case 1: return "한식"; 
-		case 2: return "중식"; 
-		case 3: return "일식/수산물"; 
-		case 4: return "분식"; 
-		case 5: return "닭/오리요리"; 
-		case 6: return "양식"; 
-		case 7: return "패스트푸드"; 
-		case 8: return "제과제빵떡케익"; 
-		case 9: return "유흥주점"; 
-		case 10: return "별식/퓨전요리"; 
-		case 11: return "커피점/카페"; 
-		case 12: return "음식배달서비스"; 
-		case 13: return "기타음식업"; 
-		case 14: return "부페"; 
-		default : 
-			System.err.println("Mypage : selectCategory 오류");
-			return "";
-		}
-	}
-	
-	
+	// 1개의 thing에 대해서 Rank 매기고 INSERT
 	public void insertRanking(Scanner keyboard) {
 		// 음식점을 선택하는 과정이 필요함 -> GUI로 구현 -> 지금은 음식점 이름이 unique하다고 가정하고 진행
-		System.out.println("랭크를 추가할 음식점(thing)의 이름을 입력하세요");
-		String thing_name = keyboard.nextLine();
-		System.out.println("Rank를 입력하세요");
-		int rank  = keyboard.nextInt();
-		keyboard.nextLine();
-		
+			
 		ThingDAO thingDAO = new ThingDAO();
 		ThingRankDAO trDAO = new ThingRankDAO();
 		int thing_id = 0;
 		try {
+			thingDAO.printTotalList();
+			System.out.println("랭크를 추가할 음식점(thing)의 이름을 입력하세요");
+			String thing_name = keyboard.nextLine();
+			System.out.println("Rank를 입력하세요");
+			int rank  = keyboard.nextInt();
+			keyboard.nextLine();
 			thing_id = thingDAO.search_id_by_name(thing_name);
 			trDAO.insert(thing_id, this.getUser_id(), rank);
 		} catch (SQLException e) {
@@ -130,12 +102,14 @@ public class Mypage extends UserPage {
 	}
 	
 	public void deleteRanking(Scanner keyboard) {
-		System.out.println("랭크를 삭제할 음식점(thing)의 이름을 입력하세요");
-		String thing_name = keyboard.nextLine();
+		
 		ThingDAO thingDAO = new ThingDAO();
 		ThingRankDAO trDAO = new ThingRankDAO();
 		int thing_id = 0;
 		try {
+			thingDAO.printTotalList();
+			System.out.println("랭크를 삭제할 음식점(thing)의 이름을 입력하세요");
+			String thing_name = keyboard.nextLine();
 			thing_id = thingDAO.search_id_by_name(thing_name);
 			trDAO.delete(thing_id, this.getUser_id());
 		} catch (SQLException e) {
@@ -147,15 +121,17 @@ public class Mypage extends UserPage {
 	}
 	
 	public void updateRanking(Scanner keyboard) {
-		System.out.println("랭크를 수정할 음식점(thing)의 이름을 입력하세요");
-		String thing_name = keyboard.nextLine();
-		System.out.println("새로운 랭크를 입력하세요");
-		int rank  = keyboard.nextInt();
-		keyboard.nextLine();
+	
 		ThingDAO thingDAO = new ThingDAO();
 		ThingRankDAO trDAO = new ThingRankDAO();
 		int thing_id = 0;
 		try {
+			thingDAO.printTotalList();
+			System.out.println("랭크를 수정할 음식점(thing)의 이름을 입력하세요");
+			String thing_name = keyboard.nextLine();
+			System.out.println("새로운 랭크를 입력하세요");
+			int rank  = keyboard.nextInt();
+			keyboard.nextLine();
 			thing_id = thingDAO.search_id_by_name(thing_name);
 			trDAO.update(thing_id, this.getUser_id(), rank);
 		} catch (SQLException e) {
@@ -165,6 +141,7 @@ public class Mypage extends UserPage {
 		trDAO = null;
 	}
 	
+	//내가 작성한 snapshot의 좋아요 수 합계를 출력
 	public void print_snapshot_like_count() {
 		Connection conn = null;
 		PreparedStatement pstmt = null;
